Stabilize worker_park_count and worker_unpark_count
